remove browser support
The pre-compiled interpreter in the browser-support folder has been out
of date for some time. Since it does not reflect the code changes
anymore I removed it for now from the master branch and created a new
branch 'browser-support'.

I could not find a way to reliably build GHCJS, neither by using NixOS
and the README's instructions nor by manually building is inside a
default Ubuntu 20.04 system. GHCJS seems to be quite a fragile project.
It seems to be necessary to use the Nix package manager or NixOS. When
it is possible to reliably build GHCJS (best: version 8.10) it should be
no problem to move the browser-support back into the master branch.